src: pomodoro: Update codestyle and refactor structure
The pomodoro.sh file doesn't follow the current project codestyle, like:
- Having a standardized parser
- Parser named 'parse_<feature-name>'
- String concatenation without brackets
- No parser error handling

Also, the structure can be improved as some refactorings can be made,
like method extractions, positioning of functions to reduce file
scrolling and refactoring of options.

This commit fixes those resulting in a more clean code. The actual size
of the file grew, but this is due to the extraction of methods and their
documentation. The 'kw pomodoro --tag' no longer shows all registered
tags, instead this functionality is provided by the '--show-tags/-s'
option. Also, the '--list/-l' option was renamed to '--check-time/-c' to
be more indicative of its behavior.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>