src: deploy: Handle command failure for statistics
In `kw deploy` feature, all commands tracked for statistics (except for
full deploy) don't handle failure for statistics purposes. The failure
of those commands is not handled at all. Either way, when a command
fails, e.g. `kw deploy --list` ssh connection fails because the target
remote machine is down, a status of `success` will be registered for
this statistic in the database (in the example,
`<ID>|list|success|<DATE>|<START_TIME>|<RUN_TIME>`).

This commit makes `kw deploy` take account of the failure of all those
commands when calling `statistics_manager`. We now capture the return
value of the effective command handlers (run_kernel_install,
modules_install, etc.) and use it to determine the `status` between
`success` and `failure`.

An interesting approach would be to pass the return value to
`statistics_manager` and let it handle it. This would move the if and
else statements from `deploy.sh` to `kwlib.sh` in the function
definition.

Also, it is worth pointing out that an `interrupted` command is not
added to the statistics and we should consider handling this case in
the future.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>