src: report: Integrate with kw_db
The `kw report` feature handles every type of report of kw tracked data.
As all data used to be stored using metadata files, `report` is
responsible for any logic related to fetching, processing, formatting,
and outputting reports. There are two points worth noting:
- The feature offers filters like by type of report (statistics or
  Pomodoro sessions) and periods of time (like a given day, week, month
  and year). The latter resulted in much of the feature's work being
  dealing with files hierarchy to fetch the proper data based on a
  period of time filter.
- Also, the file doesn't follow a pattern. For example, the logic for
  implementing the period of time filters for the Pomodoro sessions
  report is defined in `report.sh` whilst the one for the statistics
  report is defined in `statistics.sh`. The responsibilities of each
  function are not clear, as some fetch, process, and format the data
  (statistics case) and others just fetch and partly process the data
  (Pomodoro sessions case).

This commit solves the extensive logic implementation problem and lack
of standardized code by integrating `kw report` with `kw_db`.

A dedicated database and the power of SQL naturally solved the need for
complicated logic for applying period of time filters to data fetching.
This greatly reduced the complexity of the code and its size.

The integration also allowed us to enforce a pattern in the file. This
was done by:
- Unifying the function to fetch the raw data (`set_raw_data`);
- Adding a function to each type of report to process and format the
  data (`process_and_format_<statistics/pomodoro>_raw_data`);
- Unifying the output function (`show_report`).

Note:
- This change resulted in a revision of the `report` test suite that
  greatly decreased its running time.
- Just some processing functions like getting the maximum/minimum value
  from `statistics.sh` are still used. We could move them to `kwlib.sh`
  and deprecate `statistics.sh`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>