src: kw_time_and_date: Move timebox_to_sec function
The same `timebox_to_sec` function is defined in the database migration
script and in `report.sh`. Also, this function possibly will be used in
`pomodoro.sh`.

As such, this commit moves this function to the library file
`kw_time_and_date.sh` where it can be included in the files listed
above. The commit also adds documentation, error handling, and automated
tests to the function.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>