src: backup: Adapt to kw_db
When restoring a backup of kw data, if not using the `--force` flag, the
statistics and pomodoro reports, and the kernel configs data are
restored. This data is almost all the personal user data of kw (besides
the kw config files).

With the addition of `kw_db`, the file `KW_DATA_DIR/kw.db` needs to be
restored as well. Also, the legacy metadata files and directories don't
have the need to be restored and (in theory) should already be migrated
into `kw_db`, leaving just `KW_DATA_DIR/kw.db` and the actual kernel
config files to be restored.

This commit does this by not calling the functions that used to restore
the data mentioned in the first paragraph and adding and calling
functions that restore the `kw_db` file and actual kernel config files.

Note: As said before, this commit only affects the restoration without
`--force` as, in this case, the full decompressed backup is just copied
to `KW_DATA_DIR`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>