src: lib: lore: Use b4 tool for downloading patch series
The download of a patch series in `upstream-patches-ui` was done using
an in-house implementation. Although functional, the `b4` tool is better
suited for this task, as this is a dedicated tool, and its capabilities
are more vast and robust. Also, considering `b4` reliable, this results
in less code to maintain.

This commits makes `upstream-patches-ui` use `b4` to do the task of
downloading a series. The downloaded file is in a .mbx format and ready
to be applied to a git tree, i.e., the patches are in order, there are
no discussion messages and there is no cover letter. The in-house
implementation did this, but there was a file for each patch and it
considered the cover letter as a patch (which resulted in a bug).

Note: The in-house implementation is still used for fetching metadata
from the lore.kernel.org API (e.g. for checking available mailing list
and patch series). `b4` doesn't seem to handle this task, but it may be
beneficial to find a way to also delegate it.

Closes: #842

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>