src: lib: dialog_ui: Fix special chars crashing bug
In the screen creating functions of `dialog_ui.sh`, we build a dialog
command and evaluate it through `cmd_manager`. Some strings that are
concatenated in the command can contain single quotes, double quotes or
apostrophes and, because we are also enclosing the dialog's arguments in
double quotes and single quotes, crashes can happen because:
- Single/double quotes inside strings can break the arguments resulting
  in bad expansions.
- Enclosing apostrophes can be interpreted as command substitution when
  passing through `cmd_manager`.

This commit fixes this series of bugs by:
- Enforcing ANSI-C Quoting in all string interpolations, i.e., use a
  character sequence like $'<string>'. For reference, see the section
  3.1.2.4 of
  https://www.gnu.org/software/bash/manual/bash.html#Shell-Expansions.
- Escaping all single quotes of all strings that will be concatenated
  into a dialog command.

Note: A function inside `kw_string.sh` was created to handle the
escaping of single quotes and a corresponding unit test was added.
However, as we can't really mock the crashes (i.e., simulate the
creation of a dialog screen), we only check if the dialog command that
is being issued follows the two points listed above.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>