src: deploy: Ensure that deploy setup run in A/B partition
SteamOS uses a mechanism named A/B partition, where a partition (A) is
completely replaced by another similar partition (B), which makes it
easy to update devices in the user's hand. One of the affected folders
from this A/B system is the /boot directory and the directories that get
apps installed via pacman, which cause issues on kw deploy after this
update. This commit addresses this problem by moving the status file
from /opt/kw to /boot/; by doing it when the A/B partition gets updated,
the kw_status file will be removed. Finally, when kw tries to check if
the deploy was executed, it will notice that kw_status was removed, and
it will rerun kw setup.

Closes: #754

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>