src: upstream_patches_ui: Add 'Save Patches To' menu
The `upstream-patches-ui` feature relies on a configuration named
'download_to' stored in `etc/lore.config` that defines the directory
where patch files will be stored. The only way to set it is to manually
alter the file, which is not ideal as the user may input an invalid path
and needs to know beforehand where it is defined.

This commit adds a menu entry named 'Save Patches To' located inside
'Dashboard'->'Settings' that allows the user to set the configuration,
now renamed to 'save_patches_to' for accuracy, through the feature's
interface.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>