src: upstream_patches_ui: Fix 'New Patches' screen title bug
In the `upstream-patches-ui` feature, the title of the screen where new
patches from a given mailing list are shown is incorrectly displayed
when returning from a series details screen. In other words, when we
first visit the 'New Patches' screen, the title is correct (something
like 'Patches from amd-gfx)'. But, if we select a patch series to see
its details and return to the previous screen ('New Patches' screen),
the title will be wrong (something like 'Patches from 0').

This commit fixes this by adding a global variable named
`current_mailing_list` to keep track of the latest mailing list queried,
i.e., the latest mailing list that the feature has fetched new patches.
We already checked to see if `screen_sequence['RETURNING']` was set,
signaling that we are (supposedly) returning from a series details
screen and we don't need to query the mailing list again, but the list
name displayed was resetted (with `screen_sequence['SCREEN_PARAMETER']`
value) regardless. Now, the list name displayed is defined by the
`current_mailing_list` value and it is only set when not returning from
a (supposed) series details screen.

Closes: #849

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>