src: remote: fix variable name in remote2host
One of the variables passed to the rsync command had a typo, so
operations like fetching a config file from a remote machine would
not succeed. Replace `dsc` with `dst` in the rsync command.
Also, add some tests to remote2host covering both the usage of a
remote file and passing of parameters from CLI.

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Magali Lemes <magalilemes00@gmail.com>