src: upstream_patches_ui: Fix relative paths in 'Kernel Tree Path'
When selecting a Linux kernel source tree through the
`upstream-patches-ui` feature, relative paths are accepted when they
refer to a kernel tree, but the relative path is registered in the
`lore.config` file, which can break parts of the feature, depending
where it is executed.

This commit fixes this by still allowing relative paths but converting
them to absolute paths using the `realpath` command.

Closes: #865

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>