src: lib: kw_string: Add function for converting string to filename
Sometimes, we need to convert an arbitrary string into a filename.
However, there some processing may be necessary, such as:
1) Remove some bad characters
2) Replace some bad characters to change the original string the least
3) Assure that the filename isn't empty or is bigger than 255 characters

Point number 3 is based on the fact that most filesystems have a maximum
filename length of 255 characters. For reference see
https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits.

This commit adds a function to the string library for this conversion
following points 1 to 3.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>