src: upstream_patches_ui: Add 'Apply' action
This commit adds the action to apply a patchset to a given Linux kernel
tree and target branch (that can be setted in the
'Dashboard'->'Settings' screen. The patchset application updates the
target branch and creates a new one based on the target branch. The new
branch name follows the pattern:
  YYYY-MM-DD-hh-mm-ss-<patchset_title_converted_to_filename>

Under the hood the patch application is done via `git am` and can fail
for any number of reasons. In case the application fails, .rej files are
generated for the files that could not be applied by `git am` and we use
`wiggle` (if installed in the system) to try apply the rejected files.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>