src: lib: dialog_ui: Add function to create 'File Selection' screen
This commit adds a function to create a dialog screen for file
selection. Much like the directory selection screen, there are two upper
panels, the left one with the current directories and the right one with
the current files. Under the panels there is an input field where the
current path is. When hitting the 'OK' button, the path in the input
field is returned through the `menu_return_string` variable.

This commit also adds text for a help screen for the 'File Selection'
screen that can be used for with the `create_help_screen`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>