src: upstream_patches_ui: Add 'Kernel Config File' menu
When building a Linux kernel, a fundamental file is the kernel config
file, A.K.A. .config. As we are aiming to add the possibility to build
a patchset version of the kernel through the `upstream-patches-ui`
feature, it is indispensable to have a setting to contemplate this file.

This commit adds a menu entry named 'Kernel Config File' located inside
'Dashboard'->'Settings' that allows the user to set a path to a kernel
config file.

There are two options for the user to set this through the feature:
- Choosing a user .config managed by kw (added through the
  `kernel_config_manager` feature)
- Choosing an arbitrary file from the user filesystem

The first approach has the benefit of guaranteeing that a kernel config
file is chosen and the second that the .config doesn't need to be under
kw management.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>