src: upstream_patches_ui: Introduce entry file to follow coding style
The `upstream-patches-ui` feature is centralized in the
`upstream_patches_ui.sh` file, which doesn't follow the project's coding
style. For example, it doesn't have a parser function.

Moreover, the implementation of the feature is quite unique, as it is a
state-machine, in which the states are the screen-sequences in the main
loop. Because of that, the file grows uncontrollaby with functions that
aren't new capabilities, but new states.

This commit breaks the file in an entry file to the feature that follows
the kw coding style and an entry file to the state-machine.

The entry file to the feature is really barebones as it just checks if
the user is using the `-h|--help` option, calls the parser (which
doesn't accepts any option) and calls the main loop function of the
state-machine. This makes it easier to add any option to the file and
also enforces the project's coding style.

The entry file to the state-machine is defined in the directory
`src/ui/upstream_patches_ui` as `upstream_patches_ui_core.sh`. That way
we can further break down this file into other files in the same
directory to keep the code cleaner, more intuitive and easier to expand
upon.

In essence, we just moved code from a file to a new one without changing
implementation to avoid adding sneaky regressions.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>