src: ui: upstream_patches_ui: settings: Extract 'Settings' to a file
In `upstream_patches_ui_core.sh` the 'Settings' screen-sequence is
represented by the `show_settings_screen` function. It is an example of
complicated, hard to read and that does many things function.

This commit extracts the screen-sequence to a file
`src/ui/upstream_patches_ui/settings.sh`. The `show_settings_screen`
function now is broken down into modular functions. This way there is
increased readibility and allows us to test it better.

[Note]
Lacks documentation and tests for the extracted functions. Ammend this
commit with these changes.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>