src: ui: patch_hub: series_details: Unify show series functions
In the `patch-hub` feature, when showing patchset informations and
actions available, there are two different functions,
`show_series_details` and `show_bookmarked_series_details`, that are
used depending if the patchset in question was a bookmarked one or not.
The implementation of both is virtually the same and adding an action
that applies for both patchset types means duplicated code.

With that in mind, this commit unifies both functions into the
`show_series_details` function.

The `show_series_details` and `show_bookmarked_series_details` where
responsible for fetching the raw data of the patchset in question. To
make `show_series_details` more generic and clean, this responsability
is passed to the caller, as the function now accepts two arguments:
- The patchset raw data;
- A boolean indicating if the patchset is a bookmarked one or not.

Note: The boolean is only used to define which actions are available for
each case. As we plan to also unify the actions available, this argument
is sure to be removed in the future and the function shall be more
generic.

Closes: #860.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>