src: lib: lore: Revise patchset entity ID attribute
In the `kw patch-hub` feature, we make use of a database entity to
represent a patchset (a set of patches related to a patch series). An
instance of a patchset must have a minimal set of attributes, including
an ID (primary key) that uniquely identifies it. We use this entity for
showing details and running actions of patchsets in a mailing list and
in the local bookmarked database. However, the ID attribute is the
SHA256 value of all other attributes, which is an unnecessary and
unstable approach as a patchset URL in the lore.kernel.org public inbox
already uniquely identifies it. This approach also difficults
implementation, as we need to calculate the SHA256 taking care to use
only the minimal set attributes.

This commit removes the manufactured ID attribute and substitutes it for
the patchset URL in lore.kernel.org. This commit also substitutes some
functions and variables names with "series" or "patch" that references a
patchset instance to "patchset", as "series" is a vague term and "patch"
is imprecise as it may refer to a patchset with more than one patch
(these changes were only made where the ID substitution had an impact).

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>