src: ui: patch_hub: series_details: Generalize showing of patchset
This commit completely generalizes the process of showing patchset's
details and actions. Previously, there was a heavy distinction between
a bookmarked and an unbookmarked patchset in `series_details.sh` which
not only made the code less clean, but also resulted in UX problems like
having to move across menus to unbookmark a patchset that was bookmarked
by accident.

The generalization is done by making invisible the type of patchset to
`series_details.sh`, and determining the actions to take (download/
remove download/bookmark/remove bookmark) by the transitions action
statuses.

Before, only actions with a set status (asterisk inside brackets) were
accounted for, which resulted in bugs and imposed limitations. For
example, having

    [*] Download
    [*] Bookmark

and hitting the OK button resulted in the actions being run again.

Now, this redundancy can be avoid. Also, because we don't distinguish
types of patchsets, the 'Unbookmark' action was removed and it is now
done by hitting the OK button after the transition

    [*] Bookmark ---> [ ] Bookmark

which possibilitates unbookmarking a patchset that was bookmarked by
accident without switching menus. The same is valid for removing the
downloaded patchset .mbx files.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>