src: ui: patch_hub: series_details: Add documentation and tests to action handlers
In the file, there is a handler function to each possible action. Those
are really important, because, if they silently fail, for example, the
user may think the patchset has been downloaded or bookmarked where in
fact it wasn't. However, there is no documentation or unit tests for
these.

With that in mind, this commit adds documentation and tests for each of
the current 4 action handlers.

Note: References to patchsets with 'series' are changed to 'patchset'
for clarity, as mentioned in previous commits.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>