src: lib: remote: Fix ssh connection fail message with remote.config
When a ssh connection using a `remote.config` file fails, the message
shown about the failed attempt displays the wrong IP, User and Port
values, if the `remote.config` file doesn't follows this pattern for
each host:

Host <host>
  Hostname <IP address>
  Port <port number>
  User <username>

For example, if there are commented lines inbetween, or the order is
different, or there are an arbitrary number of leading whitespaces on
line, or the host has missing values (invalid host) the values displayed
on a failed ssh connection will be messed.

This happens because the implementation to build the fail message
considered that the file would strictly follow the pattern above and
picked the Hostname by the line after the 'Host' line, the Port by the
second line after the 'Host' line and the User by the third line after
the 'Host' line.

This commit fixes this bug by fetching the IP, User and Port values in
the `remote.config` file more reliably, using pattern matching with grep
instead of assuming a "well-formatted" file. It also adds an unit test
to reproduce a "bad-formatted" `remote.config` file case and an unit
test to reproduce a invalid host in the `remote.config`.

Closes: #883

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Reviewed-by: Melissa Wen <melissa.srw@gmail.com>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>