src: ui: patch_hub: patchset_details_and_actions: Make 'Download' independent
When bookmarking a patch, the 'Download' action would also be checked
(may cause confusion to the user). Also this action doesn't really have
a clear purpose, as it does the same but less than bookmarking a patch.

In this context, this commit changes the 'Download' action to the
'Download to a specific directory' action, which allows the user to
select a directory to the download on the fly.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>