src: lib: dialog_ui: Extract function to build command preamble
In most functions in the file, there is a common preamble to the dialog
command being built, that consists of the configuration to the dialog
layout (if it exists), and the box title and back title.

This commit adds the function `build_dialog_command_preamble` that
outputs this preamble as a string, and uses it to build all dialog
commands that has this common preamble.

This reduces a lot a duplicated code and gives a unit to test the build
of the preamble.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>