src: lib: dialog_ui: Extract function to run dialog command
In every function responsible to show a dialog screen (i.e., that runs
a dialog command) there is a common block of code.

This commit extracts this common block of code into a function named
`run_dialog_command` to reduce duplicated code.

Note: I don't see a way to test this function, as it manipulates file
descriptors (which messes the terminal when running tests) and the only
possible test is to assure that `cmd_manager` is actually evaluating the
right command (which is a responsability of `cmd_manager` unit tests).

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>