src: lib: lore: Add fetch of last n patchsets from mailing list
This commit implements the fetching of the last n patchsets sent to a
given public mailing list in the lore.kernel.org archives.

The global variable `NUMBER_OF_PATCHSETS_PER_FETCH` defines the number
of patchsets needed to be fetch before displayng to the user and the
global variable `TIMEFRAME_IN_DAYS` defines the size of the timeframe
considered when querying lore.kernel.org, in other words, the size of
the period or chunk of time for each request to the server.

This is a big commit that creates a lot of functions, but, essentially,
it takes the same steps that the old `processing_new_patches` function
did. This function was renamed to `fetch_latest_patchsets_from` and the
created functions are delegations within it. These delegations make
`fetch_latest_patchsets_from` somewhat of a mediator and allow us to
test each step of what used to be done entirely by
`processing_new_patches`.

The major implementation change is the addition of a while loop that
fetches patches in adjacent chunks of time until the required number of
patchsets is obtained.

It is worth noting the lore API for fetching patches. For each public
mailing list archived, there is an URL that is common to every query for
patches. For example, any query for patches of 'amd-gfx' has to have

  https://lore.kernel.org/amd-gfx/

in its URL. Query filters can be applied to match only patches that
fulfill specific conditions.

For the purposes of this commit, we need only to filter based on time.
An example of query URL to request patches of 'amd-gfx' from 4 days ago
until 2023-01-01 00:00:00 of the 'amd-gfx' public mailing list.

  https://lore.kernel.org/amd-gfx/
  ?q=rt:4.day.ago..2023-01-01T00:00:00Z+AND+NOT+s:Re&x=A

The 'rt:' refers (seemingly) to the recieved time of the message in the
lore servers. We use this filter instead of 'd:' (date time accused by
the message in the 'Date' header), because lore responds queries with an
XML ordered by the attribute filtered by 'rt:' instead of 'd:'. As we
want to have a consistent ordered list, and, preferebly avoid the
overhead of having to order the list itself, we use 'rt:'. Note that the
whole filter is 'rt:4.day.ago..2023-01-01T00:00:00Z'.

The filter 's:Re:' matches only messages without 'Re:' in the subject,
to filter out replies, as we want to have only patches.

Closes: #799

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>