src: ui: patch_hub: patch_hub_core: Fix 'Registered Mailing Lists' first access bug
When we first launch `kw patch-hub`, going to the 'Registered Mailing
Lists' screen freezes the feature. This only happens in the first launch
and it is due to the assignment of the registered mailing lists from the
config file into the global array `registered_lists` declared in
`patch_hub_core.sh`. This assignment only happens in the beginning of
the `patch_hub_main_loop` function (main loop function of the feature).

When launching the feature, if there are no lists registered in the
config file (this includes the case of the bug) the `registered_lists`
array is empty which results in the dialog command freezing.

This commit fixes this bug by assigning the registered mailing lists
from the config file into the array before the dialog command call. We
declare a new local array reference in the `registered_mailing_list`
function as the global `registered_lists` seems useless in this case.

Note: By writing this changes, I noticed that there is no reason for the
existence of the global array `registered_lists`. In a further commit,
a refactor is welcomed.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>