src: ui: patch_hub: latest_patchsets_from_mailing_list: Add pagination with incremental fetches
The `kw patch-hub` feature can display the latest N patchsets from
a given lore public mailing list to the user. However, the if the user
wants to see the (N+1)th patchset, he/she has to change the source code
(to alter the hardcoded variable that determines the number of
patchsets), launch the feature again, and fetch and process the
N patchsets that were just displayed plus the (N+1)th patchset. Also,
currently the N patchsets are displayed in a single dialog screen.

This is really impractical, so this commit adds pagination for the
fetch and processment of patchsets. Essentially, now the user can see
the N patchsets broken in many dialog screens (pages) with M patchsets
each, and, more importantly, if the user wants to see an older page from
the oldest fetched and processed page, only this (and maybe more) page
will be fetched and processed.

This change allows the user to potentially go down all the history of
patchsets from a given lore mailing list. Technically the limit of
patchsets that can be displayed is the maximum size of a Bash array, but
this data structure seems to not have a limit.

Note: Although navigating through loaded pages is fast enough, page
loadings are really slow (from many seconds to minutes) and inconsistent
accross different mailing lists. This is a combination of needed
optimizations and corrections to the code that should be addressed to
allow for a better experience to the user.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>