src: lib: dialog_ui: Make `create_menu_options` more flexible
The `create_menu_options` function is responsible to create a dialog
screen with a list of options from which the user can choose one.
However, there is no way to:

- Add a 'Help' button.
- Customize the 'Extra' button label.
- Choose only a range of options from the array passed as argument to be
  displayed.

This commit extends the function to address these points above, making
it more flexible to, for example, create a menu with pagination.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>