src: lib: dialog_ui: Add function to create rangebox screen
This commit adds a function to create rangebox screens using dialog.
A rangebox is a bar, much like a volume bar, with a minimum, maximum,
and starting value that the user can decrement or increment with the UP
and DOWN arrow keys. This allows us to guaranteedly capture an integer
from the user from a restricted range.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>