src: ui: dialog_ui: Support indexed array in `create_choice_list_screen`
The `create_choice_list_screen` displays a dialog screen with a list of
choices for the user to select one. The choices are passed as argument
as an associative array with pairs 'choice' and 'choice description'.
The problem with this is, as Bash associative arrays are hashed, there
is no way to preserve the order of definition when looping through its
elements.

As sometimes we want the order of definition of the choices to be
consistent and the description is not really necessary, this commit adds
support for indexed arrays in `create_choice_list_screen`. Indexed
arrays preserve the order of definition when looping through its
elements (in a foreach, for example), which allow us to create a choice
list with a defined display order.

This commit doesn't alters the signature of `create_choice_list_screen`,
it only allows the argument `_choices` (a Bash namereference) to be an
indexed array, instead of restricting it to an associative array, by
detecting which type was passed and taking the according action.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>