src: lib: lore: Change patchsets fetch strategy
By testing the lore API, I noticed a behaviour: lore responses to
queries are paginated and the maximum number of messages per response is
200. For example, if we make a query to a public mailing list archived
on lore and there are, say, 750 messages matched, the first response
would contain only messages 1 to 200. If we want to get the next chunk
(201 to 400) we need to add the parameter `o=200` in the query string,
which is the same as clicking on the button 'next (older)' when
accessing lore.kernel.org on the browser.

`kw patch-hub` patchset fetch strategy is to consider adjacent chunks of
time (timeframes), by making query strings to match patchsets aggainst
these chunks. However, if the number of messages matched is greater than
200, messages 201 onwards won't be present in the response, which means
that the same query with a smaller chunk of time renders the same
results.

Although, by the way the implementation was done, we don't have
a correctness problem, all the date manipulation that is made to manage
these chunks of time and keeping them consistent is unnecessary, adds
more complexity to the code, and maybe results in a performance
overhead.

With all this in mind, this commit changes the patchsets fetch strategy
by completely removing any date manipulation and moving the focus to
manipulate only the `o` parameter of lore API. The core procedures of
the strategy are mantained, because the change entails only this
"translation" from chunks of time to chunks of messages.

Most changes are contained in `src/lib/lore.sh` (the Model), but some
collateral updates are done around `kw patch-hub`, in special the
removal of the 'Lore Requests Timeframe' setting.

Note: This new strategy has a possible "unpleasant behaviour" that is,
if the list gets updated during a fetch session, patchsets may appear
duplicated, but without messing the order or any functionality. In other
words, when a new message to a given list is recieved on lore, message
"number 1" becomes "number 2" and so on, so when requesting the next
chunk of messages from lore, message 200 (that was already processed)
may be processed again as message 201. This can be considered a bug that
doesn't break the feature in any way, but can be a hassle if the list
gets updated too frequently (imagine having to race aggainst the
updates).

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>