src: lib: web: Add function to encode string for URL
A common task is to pass strings in a URL as part of a query string
(e.g. 'David Tadokoro' in '?q=David%20Tadokoro'). URLs have some
reserved characters and so, some ASCII and non-ASCII chars have to be
percent encoded (e.g. space ' ' has to be percent encoded as '%20').

In this context, add library function to encode strings for URLs, named
'url_encode'.

This function was adapted to follow kw's coding style, and the primary
reference for it is
https://github.com/dylanaraps/pure-bash-bible#percent-encode-a-string,
from the book 'pure bash bible', which is licensed under the MIT
license. Also, credits to meleu (https://github.com/meleu) who wrote a
blogpost that can be checked at https://meleu.sh/urlencode, from which
this function was first found.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>