src: ui: patch-hub: search_string_in_lore: Add screen for string query in lore
Lore API provides searching based on strings. There are filters for
searching strings in a great number of fields in the messages, like
'Subject', 'To', and so on. The API also provides a search in the whole
message, which is the focus of this commit.

In this context, add a menu in 'Dashboard' for matching messages that
contain a given string. The string is collected using an Inputbox screen
and the string is effectively matched against all messages sent to all
public mailing lists archived in lore.

Closes: #800

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>