etc/mail.config: initially add LKML to blocked-email list
`kw mail` automatically adds LKML as a receipt of all patches since this
is recommended by Kernel Doc, but Kernel Doc also requires to not send
more than 15 patches at once to the vger mailing lists and states that
the volume on that list has caused a number of developers to tune it
out. To prevent users from being surprised by this automatic addition,
initialize `mail.config` with LKML blocked. Users can restore the
default behavior once they are more familiar with the `kw mail` feature
and its options.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Tested-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Melissa Wen <melissa.srw@gmail.com>