src: deploy: Fix grub not able to update in a local deploy
When kw tries to deploy in a local mode in a Debian testing distro, it
fails due to the verification failure once the grub update is inside
sbin; it looks like that command -v does not work in that folder. This
commit addresses this issue by adding a fallback check in /usr/sbin
folder. Notice that this workaround might stop working when /usr/sbin
folder becomes restricted, but for now, it gets the job done.

Closes: #713

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>