scripts: profiler: Add tool for profiling kw executions
[What]

A "profiler" is a broad concept, but, in this context, it means software
that profiles (describes in detail) the execution of a program (a kw
feature) based on a tracing report.

[Why]

As tracing is implemented in the project, a tool that allows us to
visualize and analyze tracing reports should be handy.

A dedicated tool implemented inside the project grants the possibility
of having many ways to visualize the same tracing report, and tinker it
for our needs.

[How]

Add `scripts/profiler.sh` tool to produce a discrete profile of the flow
of execution using a tracing report produced by kw.

The tracing report directory is passed as an argument to the tool and
there are two types of profiles that can be produced.

1. `./scripts/profiler.sh --full <tracing_dir>`: Display full flow of
   execution for each thread. The output follows the format:

Thread nr. 0
--> profiler_main
  .976 milliseconds
  --> parse_profiler_options
    1.836 milliseconds
    --> kw_parse
      1.766 milliseconds
    <-- kw_parse
    1.398 milliseconds
  <-- parse_profiler_options
  1.184 milliseconds
  --> process_csv_files_for_full_profile
    213.107 milliseconds
  <-- process_csv_files_for_full_profile
  1.174 milliseconds
  --> display_full_profile
    2.208 milliseconds
  <-- display_full_profile
  .941 milliseconds
<-- profiler_main

2. `./scripts/profiler.sh --summary <tracing_dir>`: Outputs a CSV of the
   summary profile of the execution, where each entry has the name of
   the function and the total time spent on it in nanoseconds (e.g.
   'profiler_main,123456789').

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>