src: config: Fix set_config behaviour when using env
When using command like

    kw config build.arch x86_32

inside an env, kw destroys the symbolic link created by the env and
creates a new file. As a result, when switching between envs the
configuration does not persist. This commit fixes this issue by adding
the follow-symlinks option in the sed command responsible for handling
the config file.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>