tracing: tracing: Introduce lib file for enabling tracing
To make an efficient tracing of kw, we need to avoid adding any overhead
at runtime, as one of the tracing objectives is to generate a report
that can be used to analyze performance. In reality its impossible to
not add any overhead, but we can heavilly mitigate it by injecting
tracing specific code when installing/building the application.

In this sense, add a new library file to house all functions for
injecting tracing-specific code into the application. To correctly trace
the execution, we need to inject code in two places:

1) Into the main kw file (the executable `kw` in the project root)
2) Into all kw library files (almost all files inside `src`)

This change implements a function named `sync_main_kw_file_with_tracing`
that takes care of the point number 1 above.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>